+++ bin/tasks/get-block.sh
@@ -2,7 +2,7 @@
 # @Author: Roni Laukkarinen
 # @Date:   2021-11-23 18:11:41
 # @Last Modified by:   Roni Laukkarinen
-# @Last Modified time: 2022-02-08 15:52:13
+# @Last Modified time: 2022-02-08 16:03:59
 echo "${YELLOW}Checking block updates...${TXTRESET}"
 cd $HOME
 git clone https://github.com/digitoimistodude/air-blocks $BLOCKS_PATH_TEMP
@@ -52,32 +52,5 @@ else
   exit
 fi
 
-# Tasks after functions.php has been updated
-# Remove the file without any changes
-rm ${PROJECT_THEME_PATH}/functions.php
-
-# Rename the changed file to the official one
-mv ${PROJECT_THEME_PATH}/tmpfile ${PROJECT_THEME_PATH}/functions.php
-
-# Run required things after the block/<block>.sh
-# (like localization.sh that needs to be run after the <block>.sh
-# because it gets variables from that file)
-
-# Localizations
-echo "${BOLDYELLOW}Importing ACF fields (json)...${TXTRESET} "
-
-# Check if translations are needed
-if [[ ${AIR_BLOCKS_LANG} = "en" ]]; then
-  # Run localization task
-  source ${SCRIPTS_LOCATION}/tasks/localization.sh
-fi
-
-# Just import the ACF fields file
-cp -nv ${BLOCK_ACF_JSON_PATH} ${PROJECT_THEME_PATH}/acf-json/
-
-# Tasks after functions.php has been updated, again (in localization.sh)
-# Remove the file without any changes
-rm ${PROJECT_THEME_PATH}/functions.php
-
-# Rename the changed file to the official one
-mv ${PROJECT_THEME_PATH}/tmpfile ${PROJECT_THEME_PATH}/functions.php
+# Tasks that should be run after block-specific tasks
+source ${SCRIPTS_LOCATION}/tasks/post-block.sh